src/test/java/br/relesi/rlessa/servicos/LocacaoServiceTest.java
package br.relesi.rlessa.servicos;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.hamcrest.CoreMatchers.instanceOf;
import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.CoreMatchers.not;
import static org.junit.Assert.assertThat;
import static org.junit.Assert.assertTrue;

import java.util.Date;

import org.hamcrest.CoreMatchers;
import org.junit.Assert;
import org.junit.Rule;
import org.junit.Test;
import org.junit.rules.ErrorCollector;
import org.junit.rules.ExpectedException;

import br.relesi.rlessa.entidades.Filme;
import br.relesi.rlessa.entidades.Locacao;
import br.relesi.rlessa.entidades.Usuario;
import br.relesi.rlessa.utils.DataUtils;

public class LocacaoServiceTest {
	
	@Rule
	public ErrorCollector error = new ErrorCollector();
	
	@Rule
	public ExpectedException exception = ExpectedException.none();

	
	@Test
	public void teste() throws Exception {

		// cenario
		LocacaoService service = new LocacaoService();
		Usuario usuario = new Usuario("Usuario 1");
		Filme filme = new Filme("Filme 1", 2, 5.0);

		// acao
		Locacao locacao = service.alugarFilme(usuario, filme);

		
		//verificacao
		error.checkThat(locacao.getValor(), is(equalTo(5.0)));
		error.checkThat(locacao.getValor(), is(not(6.0)));
		error.checkThat(DataUtils.isMesmaData(locacao.getDataLocacao(), new Date()), is(true));
		error.checkThat(DataUtils.isMesmaData(locacao.getDataRetorno(), DataUtils.obterDataComDiferencaDias(1)), is(true));

	}
	//test elegant
	@Test(expected=Exception.class)
	public void testLocacao_filmeSemEstoque() throws Exception {
		
				// cenario
				LocacaoService service = new LocacaoService();
				Usuario usuario = new Usuario("Usuario 1");
				Filme filme = new Filme("Filme 1", 0, 5.0);

				// acao
				Locacao locacao = service.alugarFilme(usuario, filme);

	}
	//robusta
	@Test
	public void testLocacao_filmeSemEstoque_2(){
		
				// cenario
				LocacaoService service = new LocacaoService();
				Usuario usuario = new Usuario("Usuario 1");
				Filme filme = new Filme("Filme 1", 0, 5.0);

				// acao
				try {
				 service.alugarFilme(usuario, filme);
				 Assert.fail("Deveria ter lancado uma excecao");
				} catch (Exception e) {				
					assertThat(e.getMessage(), is("Filme sem estoque"));
				}

	}
	
	@Test
	public void testLocacao_filmeSemEstoque3() throws Exception {
		
				// cenario
				LocacaoService service = new LocacaoService();
				Usuario usuario = new Usuario("Usuario 1");
				Filme filme = new Filme("Filme 1", 0, 5.0);
				

				exception.expect(Exception.class);
				exception.expectMessage("Filme sem estoque");
				
				// acao
				 service.alugarFilme(usuario, filme);
	
	}
	




	
	
}
